requirements.txt
anthropic==0.42.0
supabase==2.11.0
apscheduler==3.10.4
Pillow==11.1.0
httpx==0.28.1
python-dotenv==1.0.1
python-telegram-bot==21.9
Flask==3.0.0
